Renamed utils folder as utilities for improved clarity
